Refactored code to have unified style. Empty example is now automatically added when multiple examples are provided.
