+++ samples/NSwagWithExamples/Models/Examples/Persons/Requests/PersonRequestExampleTom.cs
@@ -1,0 +1,10 @@
+using System;
+using NSwag.Examples;
+
+namespace NSwagWithExamples.Models.Examples.Persons.Requests;
+
+[ExampleAnnotation(Name = "Tom", ExampleType = ExampleType.Request)]
+public class PersonRequestExampleTom : IExampleProvider<Person>
+{
+    public Person GetExample() => new Person("Tom", "Hanks", new DateTime(1956, 7, 9));
+}
