+++ samples/NSwagWithExamples/Models/Examples/Persons/Requests/PersonTextExample1.cs
@@ -1,0 +1,9 @@
+using NSwag.Examples;
+
+namespace NSwagWithExamples.Models.Examples.Persons.Requests;
+
+[ExampleAnnotation(Name = "Search text 'inspektor'", ExampleType = ExampleType.Request)]
+public class PersonTextExample1 : IExampleProvider<string>
+{
+    public string GetExample() => "inspektor";
+}
